Refactor module names for wave and medium engines
Renamed internal modules from _medium and _wave_engine variants to spacetime_medium and spacetime_ewave for both radial_wave and spring_mass experiments. Updated all relevant import statements to reflect the new module names, improving clarity and consistency in module naming.
