+++ openwave/xperiments/_archives/spring_mass/spacetime_ewave_euler.py
